Modified shoot position, spline to ring pickup
Originally had errors from changing the shoot angle,
so adjusted it's lateral position instead.
This means the shoot and park positions are not 
vertically aligned with the ring pickup position,
so a spline is used to grab the rings.

That said, the bouncing at the start and end of 
trajectories seemed to be due to the tangent 
value being in the wrong direction...
with this fixed, angle adjustments might work.